FIX: Rename package out of fmu namespace
