+++ tests/test_api/test_api_v1_main.py
@@ -3,8 +3,8 @@
 from fastapi import status
 from fastapi.testclient import TestClient
 
-from fmu.settings.api.__main__ import app
-from fmu.settings.api.config import settings
+from fmu_settings_api.__main__ import app
+from fmu_settings_api.config import settings
 
 client = TestClient(app)
